// packages/unit/src/metadata.d.ts
export interface TestDecorator<T extends TestMetadata> {
    (timeout: number): MethodDecorator;
    (metadata?: T): MethodDecorator;
}
// ...
export interface TestCase extends TestDecorator<TestCaseMetadata> {
    /**
     * @Test decorator. define the method of class as unit test case.  Describe a specification or test-case with the given `title` and callback `fn` acting
     * as a thunk.
     *
     * @param {string} title test case title.
     * @param {number} [timeout] test case timeout.
     * @param {number} [setp] test case setp order in this test suite.
     */
    (title?: string, timeout?: number, setp?: number): MethodDecorator;
}
/**
 * @Test decorator. define the method of class as unit test case.  Describe a specification or test-case with the given `title` and callback `fn` acting
 * as a thunk.
 *
 * @export
 * @interface TestCase
 */
export declare const Test: TestCase;
/**
 * @BeforeAll decorator. define the method of class as unit test action run before all test case.
 *
 * @export
 * @interface IBeforeTestDecorator
 * @extends {TestDecorator<TestMetadata>}
 */
export interface BeforeAll extends TestDecorator<TestMetadata> {
}
/**
 * @BeforeAll decorator. define the method of class as unit test action run before all test case.
 *
 * @export
 * @interface BeforeAll
 */
export declare const BeforeAll: BeforeAll;
/**
 * @BeforeAll decorator. define the method of class as unit test action run before all test case.
 *
 * @export
 * @interface BeforeAll
 */
export declare const Before: BeforeAll;
/**
 * @BeforeEach decorator. define the method of class as unit test action run before each test case.
 *
 * @export
 * @interface BeforeEach
 * @extends {TestDecorator<TestMetadata>}
 */
export interface BeforeEach extends TestDecorator<TestMetadata> {
}
/**
 * @BeforeEach decorator. define the method of class as unit test action run before each test case.
 *
 * @export
 * @interface BeforeEach
 */
export declare const BeforeEach: BeforeEach;
/**
 * @AfterAll decorator. define the method of class as unit test action run after all test case.
 *
 * @export
 * @interface AfterAll
 * @extends {TestDecorator<TestMetadata>}
 */
export interface AfterAll extends TestDecorator<TestMetadata> {
}
/**
 * @AfterAll decorator. define the method of class as unit test action run after all test case.
 *
 * @export
 * @interface AfterAll
 */
export declare const AfterAll: AfterAll;
/**
 * @AfterAll decorator. define the method of class as unit test action run after all test case.
 *
 * @export
 * @interface AfterAll
 */
export declare const After: AfterAll;
/**
 * @AfterEach decorator. define the method of class as unit test action run after each test case.
 *
 * @export
 * @interface AfterEach
 */
export interface AfterEach extends TestDecorator<TestMetadata> {
}
/**
 * @AfterEach decorator. define the method of class as unit test action run after each test case.
 *
 * @export
 * @interface AfterEach
 */
export declare const AfterEach: AfterEach;
export interface SuiteDef extends TypeDef {
    suite?: boolean;
    e2e?: boolean;
}
/**
 * Test metadata.
 *
 * @export
 * @interface TestMetadata
 * @extends {MethodMetadata}
 */
export interface TestMetadata {
    /**
     * test action type.
     *
     * @type {string}
     */
    action?: string;
    /**
     * test timeout.
     *
     * @type {number}
     */
    timeout?: number;
}
/**
 * before test metadta.
 *
 * @export
 * @interface BeforeTestMetadata
 * @extends {TestMetadata}
 */
export interface BeforeTestMetadata extends TestMetadata {
}
/**
 * before each test metadta.
 *
 * @export
 * @interface BeforeEachTestMetadata
 * @extends {TestMetadata}
 */
export interface BeforeEachTestMetadata extends TestMetadata {
}
/**
 * test case metadata.
 *
 * @export
 * @interface TestCaseMetadata
 * @extends {TestMetadata}
 */
export interface TestCaseMetadata extends TestMetadata {
    /**
    * test case title.  Describe a specification or test-case with the given `title` and callback `fn` acting
    * as a thunk.
    *
    * @type {string}
    */
    title?: string;
    /**
     * test setp
     *
     * @type {number}
     */
    setp?: number;
}
/**
 * Suite metadata.
 *
 * @export
 * @interface SuiteMetadata
 * @extends {AnnotationMetadata}
 */
export interface SuiteMetadata extends AnnotationMetadata {
    /**
     * test suite describe message.
     *
     * @type {string}
     */
    describe?: string;
    /**
     * all action default timeout.
     *
     * @type {number}
     */
    timeout?: number;
    /**
     * suite define.
     *
     * @type {boolean}
     */
    suite?: boolean;
    /**
     * Mark this as an E2E test suite.
     * 标记为E2E测试套件
     *
     * @type {boolean}
     */
    e2e?: boolean;
}